# backend/contract/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Code-first Filebonsai API. The PostgreSQL profile authenticates the local owner and derives catalog scope from server-side membership.",
    "title" : "Filebonsai Catalog",
    "version" : "0.1.0"
  },
  "servers" : [ {
    "url" : "http://127.0.0.1:8080"
  } ],
  "security" : [ {
    "sessionCookie" : [ ]
  } ],
  "paths" : {
...
    "/api/v1/uploads" : {
      "post" : {
        "description" : "The Idempotency-Key UUID is scoped to the authenticated workspace and principal.",
        "operationId" : "beginUpload",
        "parameters" : [ {
          "in" : "header",
          "name" : "Idempotency-Key",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BeginUploadRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                }
              }
            },
            "description" : "Created or replayed durable upload intent",
            "headers" : {
              "Location" : {
                "schema" : {
                  "type" : "string"
                },
                "style" : "simple"
              }
            }
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_REQUEST or VALIDATION_FAILED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "AUTH_REQUIRED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CSRF_INVALID for state-changing requests"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UPLOAD_NOT_FOUND or ENTRY_NOT_FOUND, including inaccessible resources"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "State, name, idempotency, size, or digest conflict"
          },
          "410" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "EXPIRED"
          },
          "413" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "TOO_LARGE"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "STORAGE_UNAVAILABLE",
            "headers" : {
              "Retry-After" : {
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          }
        },
        "summary" : "Reserve a new file upload",
        "tags" : [ "Transfers" ]
      }
    },
    "/api/v1/uploads/{id}" : {
      "delete" : {
        "operationId" : "cancelUpload",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                }
              }
            },
            "description" : "Cancelled upload state"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_REQUEST or VALIDATION_FAILED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "AUTH_REQUIRED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CSRF_INVALID for state-changing requests"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UPLOAD_NOT_FOUND or ENTRY_NOT_FOUND, including inaccessible resources"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "State, name, idempotency, size, or digest conflict"
          },
          "410" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "EXPIRED"
          },
          "413" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "TOO_LARGE"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "STORAGE_UNAVAILABLE",
            "headers" : {
              "Retry-After" : {
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          }
        },
        "summary" : "Cancel before finalization starts",
        "tags" : [ "Transfers" ]
      },
      "get" : {
        "operationId" : "getUpload",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                }
              }
            },
            "description" : "Durable upload state"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_REQUEST or VALIDATION_FAILED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "AUTH_REQUIRED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CSRF_INVALID for state-changing requests"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UPLOAD_NOT_FOUND or ENTRY_NOT_FOUND, including inaccessible resources"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "State, name, idempotency, size, or digest conflict"
          },
          "410" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "EXPIRED"
          },
          "413" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "TOO_LARGE"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "STORAGE_UNAVAILABLE",
            "headers" : {
              "Retry-After" : {
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          }
        },
        "summary" : "Read durable upload state",
        "tags" : [ "Transfers" ]
      }
    },
    "/api/v1/uploads/{id}/complete" : {
      "post" : {
        "operationId" : "completeUpload",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                }
              }
            },
            "description" : "Available or reconciling upload state"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_REQUEST or VALIDATION_FAILED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "AUTH_REQUIRED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CSRF_INVALID for state-changing requests"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UPLOAD_NOT_FOUND or ENTRY_NOT_FOUND, including inaccessible resources"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "State, name, idempotency, size, or digest conflict"
          },
          "410" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "EXPIRED"
          },
          "413" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "TOO_LARGE"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "STORAGE_UNAVAILABLE",
            "headers" : {
              "Retry-After" : {
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          }
        },
        "summary" : "Publish verified staged bytes",
        "tags" : [ "Transfers" ]
      }
    },
    "/api/v1/uploads/{id}/content" : {
      "put" : {
        "description" : "Every retry sends the complete body again from byte zero.",
        "operationId" : "putUploadContent",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "in" : "header",
          "name" : "X-CSRF-TOKEN",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/octet-stream" : {
              "schema" : {
                "type" : "string",
                "format" : "binary"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UploadResponse"
                }
              }
            },
            "description" : "Verified staged upload state"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "INVALID_REQUEST or VALIDATION_FAILED"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "AUTH_REQUIRED"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "CSRF_INVALID for state-changing requests"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "UPLOAD_NOT_FOUND or ENTRY_NOT_FOUND, including inaccessible resources"
          },
          "409" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "State, name, idempotency, size, or digest conflict"
          },
          "410" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "EXPIRED"
          },
          "413" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "TOO_LARGE"
          },
          "503" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description" : "STORAGE_UNAVAILABLE",
            "headers" : {
              "Retry-After" : {
                "schema" : {
                  "type" : "integer"
                },
                "style" : "simple"
              }
            }
          }
        },
        "summary" : "Stream a whole upload body",
        "tags" : [ "Transfers" ]
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "ApiErrorResponse" : {
        "required" : [ "code", "fieldErrors", "message", "requestId", "status" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Open string, not an enum. Known codes are documented per operation. Fall back to status and message for unknown codes."
          },
          "fieldErrors" : {
            "type" : "array",
            "description" : "Empty when no individual field is responsible",
            "items" : {
              "$ref" : "#/components/schemas/FieldErrorResponse"
            }
          },
          "message" : {
            "type" : "string"
          },
          "requestId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "BeginUploadRequest" : {
        "required" : [ "name", "parentId", "sizeBytes" ],
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "description" : "Logical filename; never used as a storage path"
          },
          "parentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sha256" : {
            "pattern" : "^[0-9a-fA-F]{64}$",
            "type" : "string",
            "description" : "Optional expected SHA-256 as 64 hexadecimal characters",
            "nullable" : true
          },
          "sizeBytes" : {
            "pattern" : "0|[1-9][0-9]*",
            "type" : "string",
            "description" : "Exact decimal byte count, up to the configured upload limit"
          }
        }
      },
...
      "FieldErrorResponse" : {
        "required" : [ "code", "field", "message" ],
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string",
            "description" : "Open string code; clients must allow unknown codes"
          },
          "field" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "description" : "Empty when no individual field is responsible"
      },
...
      "UploadResponse" : {
        "required" : [ "computedSha256", "entryId", "expectedSha256", "expiresAt", "id", "name", "parentId", "sizeBytes", "state", "versionId" ],
        "type" : "object",
        "properties" : {
          "computedSha256" : {
            "type" : "string",
            "nullable" : true
          },
          "entryId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "expectedSha256" : {
            "type" : "string",
            "nullable" : true
          },
          "expiresAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "parentId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "sizeBytes" : {
            "pattern" : "0|[1-9][0-9]*",
            "type" : "string"
          },
          "state" : {
            "type" : "string",
            "enum" : [ "INITIATED", "RECEIVING", "STAGED", "FINALIZING", "RECONCILING", "AVAILABLE", "CANCELLED", "EXPIRED", "FAILED" ]
          },
          "versionId" : {
            "type" : "string",
            "format" : "uuid"
          }
        }
      }
    },
    "securitySchemes" : {
      "sessionCookie" : {
        "in" : "cookie",
        "name" : "FILEBONSAI_SESSION",
        "type" : "apiKey"
      }
    }
  }
}